Fix the frame buffer usage
The original logic would stop if the number of frames in the buffer
smaller than the threshold, which means this buffer only shifts these
frames rather than playing the buffering role correctly.

The "isReady" variable would use the frames in the frame buffer
correctly and stop when short of frame.